refactor: extract JobPollResult + GroupPollResult into Runner/PollResults.swift
Part of phase 1 (#756) — RunnerBarCore prep.
Pure code move, RunnerStore extensions remain in RunnerPollState.swift.
